refactor(android): decompose AndroidPeripheral 876->334 loc via extension functions
